Switch to JSON configuration
As reported in #6, when generating UCL configuration some configuration
cannot be set up by the module.  It also generate a large amount of
resources (more than 100 in my considered simple configuration).

Switch to generating JON files instead of UCL ones, as this is also
supported by Rspamd and allows the end-user to have a direct mapping
between Puppet data structures and what is generated in that JON files.

As UCL manipulation code is not needed anymore, remove it.

While this is a backwards incompatible change, for users who to not
manipulate the module "internal" classes no change is needed.  In my
case, the module now manage 25 resources.  Users of more complex
configuration are likely to experience even better resources reduction.